Add a test mode
When enabled, DeliveryBoy::Fake will handle all operations, which will
be applied against a fake, in-memory message store.

During normal operations, an instance of Instance will handle
everything.
